feat: Add Card component to Intro.tsx
Added a new Card component to display information about SITCON and the Code of Conduct. The Card component includes an icon, title, description, and a link to learn more. This feature will enhance the user experience and provide more information to the users. The commit includes changes to the Intro.tsx file and the addition of two SVG icons for the cards.